requirements.txt
fastapi
uvicorn[standard]
requests
redis
aiohttp
python-dotenv

